+++ account/views.py
@@ -17,6 +17,11 @@ from django.template import loader
 
 from django.core.mail import EmailMultiAlternatives
 
+from django.contrib.auth.mixins import UserPassesTestMixin
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.contrib.auth.decorators import login_required
+
+from django.contrib.auth import update_session_auth_hash
 from django.contrib.auth import (
     REDIRECT_FIELD_NAME, get_user_model, login as auth_login,
     logout as auth_logout, update_session_auth_hash,
@@ -32,7 +37,11 @@ from django.contrib.auth.views import (
 )
 
 from django.utils import timezone
+
 from django.utils.decorators import method_decorator
+from django.views.decorators.debug import sensitive_post_parameters
+from django.views.decorators.csrf import csrf_protect
+
 from django.utils.translation import gettext, gettext_lazy as _
 from django.utils.http import (
     url_has_allowed_host_and_scheme, urlsafe_base64_decode,
@@ -41,6 +50,7 @@ from django.utils.http import (
 from django.views.decorators.cache import never_cache
 
 from django.views.generic.base import View
+from django.views.generic.edit import FormView
 
 from django.contrib import messages
 
@@ -49,7 +59,8 @@ from .forms import (
     AuthenticationForm,
     PasswordResetForm,
     SetPasswordForm,
-    UserCreationForm
+    UserCreationForm,
+    PasswordChangeForm
 )
 
 # Third party modules
@@ -62,8 +73,22 @@ __all__ = [
     "CustomPasswordResetDoneView",
     "CustomPasswordResetConfirmView",
     "CustomPasswordResetCompleteView",
-    "CreateUserView"
+    "CreateUserView",
+    "UserDashboard",
+    "UserPasswordChangeView"
 ]
+
+class PasswordContextMixin:
+    extra_context = None
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context.update({
+            'title': self.title,
+            **(self.extra_context or {})
+        })
+        return context
+
 
 class CreateUserView(View):
     """Class for creating user with no priviledges"""
@@ -108,6 +133,8 @@ class CreateUserView(View):
         }
         return render(request=self.request, template_name=self.template_name, context=self.get_context_data())
 
+    @method_decorator(sensitive_post_parameters())
+    @method_decorator(csrf_protect)
     def post(self, *args, **kwargs):
         form = self.get_form_class(self.request.POST)
         if form.is_valid():
@@ -247,3 +274,59 @@ class CustomPasswordResetCompleteView(PasswordResetCompleteView):
         context = super().get_context_data(**kwargs)
         context["login_url"] = resolve_url(settings.LOGIN_URL)
         return context
+
+
+class UserDashboard(LoginRequiredMixin, UserPassesTestMixin, View):
+    """
+    UserDashboard where only the user can access
+    """
+    template_name = "account/dashboard/index.html"
+
+    def get_context_data(self, **kwargs):
+        context = {
+            "title": _("CodeTopia | %s"% self.kwargs.get("username"))
+        }
+        return context
+
+    def test_func(self):
+        """
+        Test the request sender if he has permission to access the page
+        """
+        return self.request.user.username == self.kwargs.get("username")
+
+    def get_test_func(self):
+        """
+        Call all tests that has to be passed to access the page
+        """
+        return self.test_func
+
+    def get(self, *args, **kwargs):
+        self.get_context_data().update({
+
+        })
+        return render(request=self.request, template_name=self.template_name, context=self.get_context_data())
+
+
+class UserPasswordChangeView(PasswordContextMixin, FormView):
+    form_class = PasswordChangeForm
+    success_url = reverse_lazy('password_change_done')
+    template_name = "account/dashboard/change_password.html"
+    title = _("CodeTopia | Chnage Password")
+
+    @method_decorator(sensitive_post_parameters())
+    @method_decorator(csrf_protect)
+    @method_decorator(login_required)
+    def dispatch(self, *args, **kwargs):
+        return super().dispatch(*args, **kwargs)
+
+    def get_form_kwargs(self):
+        kwargs = super().get_form_kwargs()
+        kwargs['user'] = self.request.user
+        return kwargs
+
+    def form_valid(self, form):
+        form.save()
+        # Updating the password logs out all other sessions for the user
+        # except the current one.
+        update_session_auth_hash(self.request, form.user)
+        return super().form_valid(form)